+++ frontend/src/pages/SongDetailPage.tsx
@@ -1,49 +1,135 @@
+import { useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
+import CommentList from '@/features/comments/components/CommentList';
+import KillingPartInfo from '@/features/songs/components/KillingPartInfo';
 import Thumbnail from '@/features/songs/components/Thumbnail';
-import VoteInterface from '@/features/songs/components/VoteInterface';
-import { VoteInterfaceProvider } from '@/features/songs/components/VoteInterfaceProvider';
 import { useGetSongDetail } from '@/features/songs/remotes/useGetSongDetail';
-import { VideoPlayerProvider } from '@/features/youtube/components/VideoPlayerProvider';
 import Youtube from '@/features/youtube/components/Youtube';
+import useVideoPlayerContext from '@/features/youtube/hooks/useVideoPlayerContext';
+import Flex from '@/shared/components/Flex';
+import Spacing from '@/shared/components/Spacing';
+import SRAlert from '@/shared/components/SRAlert';
+import SRHeading from '@/shared/components/SRHeading';
+import ToggleGroup from '@/shared/components/ToggleGroup/ToggleGroup';
+import ToggleSwitch from '@/shared/components/ToggleSwitch/ToggleSwitch';
 
 const SongDetailPage = () => {
-  const { id: songIdParam } = useParams();
-  const { songDetail } = useGetSongDetail(Number(songIdParam));
+  const { id = '' } = useParams();
+  const [isRepeat, setIsRepeat] = useState(true);
+  const [killingRank, setKillingRank] = useState<number | null>(null);
+  const { songDetail } = useGetSongDetail(Number(id));
+  const timer = useRef<number>(-1);
+  const { videoPlayer } = useVideoPlayerContext();
+
+  useEffect(() => {
+    if (!videoPlayer?.seekTo) return;
+    if (!songDetail) return;
+
+    const part = songDetail.killingParts?.find((part) => part.rank === killingRank);
+
+    if (!part) {
+      videoPlayer?.seekTo(0, true);
+      videoPlayer?.playVideo();
+      return;
+    } else {
+      videoPlayer?.seekTo(part.start, true);
+    }
+
+    videoPlayer?.playVideo();
+
+    if (isRepeat) {
+      timer.current = window.setInterval(
+        () => {
+          videoPlayer?.seekTo(part.start, true);
+        },
+        (part.end - part.start) * 1000
+      );
+    }
+
+    return () => {
+      window.clearInterval(timer.current);
+    };
+  }, [isRepeat, killingRank, videoPlayer, songDetail]);
 
   if (!songDetail) return;
-  const { id, title, singer, videoLength, songVideoUrl, albumCoverUrl } = songDetail;
-  // TODO: videoId 자체가 응답값으로 오도록 API 협의
-  // TODO: Jacket img src API 추가 협의
+  const { killingParts, singer, title, songVideoUrl, albumCoverUrl } = songDetail;
+  const killingPart = killingParts?.find((part) => part.rank === killingRank);
+
   const videoId = songVideoUrl.replace('https://youtu.be/', '');
 
+  const changeKillingRank = (rank: number) => {
+    setKillingRank(rank);
+  };
+
+  const toggleRepetition = () => {
+    setIsRepeat(!isRepeat);
+  };
+
   return (
-    <Container>
-      <BigTitle>킬링파트 투표 🔖</BigTitle>
+    <Wrapper>
+      <SRHeading>킬링파트 듣기 페이지</SRHeading>
+      <BigTitle aria-label="킬링파트 듣기">킬링파트 듣기 🎧</BigTitle>
+      <Spacing direction="vertical" size={20} />
       <SongInfoContainer>
-        <Thumbnail src={albumCoverUrl} alt={`${title} 앨범 자켓`} />
+        <Thumbnail src={albumCoverUrl} />
         <Info>
-          <SongTitle>{title}</SongTitle>
-          <Singer>{singer}</Singer>
+          <SongTitle aria-label={`노래 ${title}`}>{title}</SongTitle>
+          <Singer aria-label={`가수 ${singer}`}>{singer}</Singer>
         </Info>
       </SongInfoContainer>
-      <VideoPlayerProvider>
-        <Youtube videoId={videoId} />
-        <VoteInterfaceProvider>
-          <VoteInterface videoLength={videoLength} songId={id} />
-        </VoteInterfaceProvider>
-      </VideoPlayerProvider>
-    </Container>
+      <Spacing direction="vertical" size={20} />
+      <Youtube videoId={videoId} />
+      <Spacing direction="vertical" size={16} />
+      <RegisterTitle aria-label="인기 많은 킬링파트를 들어보세요">
+        인기 많은 킬링파트를 들어보세요 🎧
+      </RegisterTitle>
+      <Spacing direction="vertical" size={16} />
+      <ToggleGroup onChangeButton={changeKillingRank}>
+        <ToggleGroup.Button tabIndex={0} index={1} aria-label="1등 킬링파트 노래 듣기">
+          1st
+        </ToggleGroup.Button>
+        <Spacing direction="horizontal" size={10} />
+        <ToggleGroup.Button tabIndex={0} index={2} aria-label="2등 킬링파트 노래 듣기">
+          2nd
+        </ToggleGroup.Button>
+        <Spacing direction="horizontal" size={10} />
+        <ToggleGroup.Button index={3} aria-label="3등 킬링파트 노래 듣기">
+          3rd
+        </ToggleGroup.Button>
+        <Spacing direction="horizontal" size={10} />
+        <ToggleGroup.Button index={4} aria-label="노래 전체 듣기">
+          전체
+        </ToggleGroup.Button>
+      </ToggleGroup>
+      <Spacing direction="vertical" size={10} />
+      <SwitchWrapper>
+        <SwitchLabel htmlFor="repetition">반복재생</SwitchLabel>
+        <ToggleSwitch
+          id="repetition"
+          on={toggleRepetition}
+          off={toggleRepetition}
+          defaultToggle={isRepeat}
+        />
+      </SwitchWrapper>
+      <Spacing direction="vertical" size={10} />
+      <KillingPartInfo killingPart={killingPart} />
+      <Spacing direction="vertical" size={10} />
+      {killingPart && <CommentList songId={id} partId={killingParts[killingRank! - 1].id} />}
+      <SRAlert>{`${killingRank === 4 ? '전체' : `${killingRank}등 킬링파트`} 재생`}</SRAlert>
+    </Wrapper>
   );
 };
 
 export default SongDetailPage;
 
-const Container = styled.section`
-  display: flex;
-  width: 100%;
+const BigTitle = styled.h2`
+  font-size: 28px;
+  font-weight: 700;
+`;
+
+const Wrapper = styled(Flex)`
   flex-direction: column;
-  gap: 20px;
 `;
 
 const SongInfoContainer = styled.div`
@@ -74,7 +160,24 @@ const Singer = styled.p`
   }
 `;
 
-const BigTitle = styled.h2`
-  font-size: 28px;
+const SwitchWrapper = styled.div`
+  display: flex;
+  justify-content: end;
+  margin: 0 8px;
+  column-gap: 8px;
+`;
+
+const SwitchLabel = styled.label`
+  color: ${({ theme: { color } }) => color.white};
+  font-size: 14px;
+`;
+
+const RegisterTitle = styled.p`
+  font-size: 22px;
   font-weight: 700;
+  color: ${({ theme: { color } }) => color.white};
+
+  @media (max-width: ${({ theme }) => theme.breakPoints.md}) {
+    font-size: 18px;
+  }
 `;
