+++ frontend/src/router.tsx
@@ -1,28 +1,29 @@
 import { createBrowserRouter } from 'react-router-dom';
 import { VideoPlayerProvider } from './features/youtube/components/VideoPlayerProvider';
+import MainPage from './pages/MainPage';
+import PartCollectingPage from './pages/PartCollectingPage';
 import SongDetailPage from './pages/SongDetailPage';
-import SongPage from './pages/SongPage';
-import SongPopularPage from './pages/SongPopularPage';
 import Layout from './shared/components/Layout/Layout';
+import ROUTE_PATH from './shared/constants/path';
 
 const router = createBrowserRouter([
   {
-    path: '/',
+    path: ROUTE_PATH.ROOT,
     element: <Layout />,
     children: [
       {
         index: true,
-        element: <SongPopularPage />,
+        element: <MainPage />,
       },
       {
-        path: '/:id',
-        element: <SongDetailPage />,
+        path: `${ROUTE_PATH.COLLECT}/:id`,
+        element: <PartCollectingPage />,
       },
       {
-        path: '/song/:id',
+        path: `${ROUTE_PATH.SONG_DETAIL}/:id`,
         element: (
           <VideoPlayerProvider>
-            <SongPage />
+            <SongDetailPage />
           </VideoPlayerProvider>
         ),
       },
