+++ frontend/src/shared/constants/path.ts
@@ -1,0 +1,7 @@
+const ROUTE_PATH = {
+  ROOT: '/',
+  COLLECT: 'collect',
+  SONG_DETAIL: 'songs',
+} as const;
+
+export default ROUTE_PATH;
